requirements.txt
fastapi==0.115.6
httpx==0.28.1
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
aiosqlite==0.20.0
bcrypt==4.2.1
cloud-sql-python-connector[asyncpg]==1.13.0
google-genai==1.72.0
google-cloud-aiplatform==1.147.0
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
